Add feedback changes #1266
