+++ analysis/model/src/main/kotlin/de/maibornwolff/codecharta/progresstracker/ParsingUnit.kt
@@ -1,0 +1,5 @@
+package de.maibornwolff.codecharta.progresstracker
+
+enum class ParsingUnit{
+    Byte, Files
+}
